refactor: directory structure
