+++ src/controller/tag/response.ts
@@ -1,30 +1,5 @@
 import * as schemaHelper from "../../pkg/schemaHelper";
 
-// type TagBaseResponse = {
-//   id: string;
-//   name: string;
-//   color: string;
-//   created_at: Date;
-//   updated_at: Date;
-// };
-
-// export type CreateTagResponse = TagBaseResponse & {
-//   status: string;
-// };
-
-// export type UpdateTagResponse = TagBaseResponse & {
-//   status: string;
-// };
-
-// export type ListTagsResponse = {
-//   status: string;
-//   tags: TagBaseResponse[];
-// };
-
-// export type DeleteTagResponse = {
-//   status: string;
-// };
-
 export type CreateTagResponse = schemaHelper.ResponseData<"/tags", "post">;
 
 export type UpdateTagResponse = schemaHelper.ResponseData<"/tags/:id", "patch">;
