+++ src/route/middlewares/auth.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { verifyToken } from "../../core/jwt";
+import { verifyToken } from "../../pkg/jwt";
 
 export const auth = (req: Request, res: Response, next: NextFunction) => {
   try {
