+++ src/route/user.ts
@@ -8,7 +8,7 @@ import {
   updateUserController,
   deleteUserController,
 } from "../controller/user/user_controller";
-import { validate } from "../pkg/validate";
+import { validate } from "./utils/validate";
 import { auth } from "./middlewares/auth";
 
 const userRouter = Router();
