+++ src/service/badge/check-achievements/level_check_service.ts
@@ -1,7 +1,7 @@
 import { PrismaClientWithTx } from "../../../db/types";
 import { BadgeModel } from "../../../model/badge/badge_model";
 import { BadgeDetailModel } from "../../../model/badgeDetail/badge_detail_model";
-import { HttpError } from "../../../pkg/httpError";
+import { HttpError } from "../../../utils/httpError";
 
 const features = {
   stage1: {
