+++ src/service/quest/force_finish_quest_service.ts
@@ -1,7 +1,7 @@
 import { prisma } from "../../db/db";
 import { QuestModel } from "../../model/quest/quest_model";
 import { WeeklyReportModel } from "../../model/weeklyReport/weekly_report_model";
-import { HttpError } from "../../pkg/httpError";
+import { HttpError } from "../../utils/httpError";
 
 type InputDTO = { id: string };
