+++ src/service/weeklyReport/get_weekly_report_feedback_service.ts
@@ -1,5 +1,5 @@
 import { WeeklyReportModel } from "../../model/weeklyReport/weekly_report_model";
-import { HttpError } from "../../pkg/httpError";
+import { HttpError } from "../../utils/httpError";
 
 export const getWeeklyReportFeedBackService = async (inputDTO: {
   weeklyReportId: string;
