+++ src/service/weeklyReport/list_weekly_reports_service.ts
@@ -1,5 +1,5 @@
 import { WeeklyReportModel } from "../../model/weeklyReport/weekly_report_model";
-import { HttpError } from "../../pkg/httpError";
+import { HttpError } from "../../utils/httpError";
 
 export const listWeeklyReportsService = async (inputDTO: {
   userId: string;
